+++ src/app/layout.tsx
@@ -3,9 +3,8 @@ import { Suspense } from "react";
 import { getDocument } from "../db/db";
 import { Config } from "../lib/config-provider";
 import { cn } from "@/lib/utils";
-import Navbar from "@/components/navbar/navbar";
 import {Providers} from "./providers";
-import NewNavbar from "@/components/newnavbar/navbar";
+import NewNavbar from "@/components/navbar/navbar";
 import { Footer } from "@/components/footer/footer";
 import "./globals.css";
 import Loading from "./loading";
