+++ src/components/navbar/navbar.tsx
@@ -1,128 +1,99 @@
 "use client";
 
-import { useState } from "react";
-import { NavItem } from "./nav-item";
-import { NavigationMenu } from "../navigation-menu/navigation-menu";
-import Icon from "../icons/icon";
-import Accordion from "../accordion/accordion";
-import {
-  usePathname as getPathname,
-} from "next/navigation";
-import type { Link } from "@/lib/linktype";
-import type { Config } from "@/lib/config-provider";
+import { Config } from "@/lib/config-provider";
+import { Navbar, NavbarBrand, NavbarContent, NavbarItem, NavbarMenu, NavbarMenuItem, NavbarMenuToggle } from "@nextui-org/react";
+import Link from "next/link";
+import { usePathname as getPathname } from "next/navigation";
+import type { Link as LinkType } from "@/lib/linktype";
+import Logo from "../logo/logo";
+import Dropdown from "../navigation-menu/navigation-menu";
+import { Fragment, useState } from "react";
 
-import "./navbar.css";
-
-export function getActiveClassProps(url: string) {
+function isCurrentPath(url: string) {
   const pathnameArr: string[] = getPathname().split("/");
   const urlArr: string[] = url.split("/");
-  if (pathnameArr[1] == urlArr[1]) {
-    return {
-      className: "rounded underline decoration-2 font-bold text-black underline-offset-8",
-      "aria-label": `${urlArr[1]} current page`,
-    };
-  } else {
-    return {
-      className: "rounded hover:underline hover:text-slate-500 underline-offset-8",
-    };
-  }
+  return pathnameArr[1] == urlArr[1];
 }
 
-export default function Navbar({
-  nav,
+function getActiveClassProps(isActive: boolean) {
+  const baseClass: string = "rounded underline-offset-8 text-xl font-normal";
+  return {
+    className: baseClass + (isActive ? "underline decoration-2 font-bold text-black" : "rounded hover:underline hover:text-slate-500"),
+  };
+}
+
+export default function NewNavbar({
+  data,
+  title,
+  description,
 }: Readonly<{
-  nav: Config;
+  data: Config;
+  title: string;
+  description: string;
 }>) {
-  const [isNavOpen, setNavOpen] = useState(false);
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
 
-  if(isNavOpen){
-    document.body.style.overflow = 'hidden';
-  }
-
-  // Mobile Navbar
-  if (isNavOpen) {
-    return (
-      <>
-        <button
-          className="cursor-pointer z-20 pr-4 text-gray-500 md:hidden"
-          onClick={() => setNavOpen(!isNavOpen)}
-
-        >
-          <Icon type="fas-x" className="w-8" altText="Close navigation menu" />
-        </button>
-
-        <ol className="flex z-10 flex-col justify-center items-center absolute top-0 left-0 w-full h-screen bg-white">
-          {nav?.map((link: Link, i: number) => (
-            <li
-              key={i}
-              className="py-4 cursor-pointer capitalize text-4xl hover:text-blue-900"
-            >
-              {link.type === "link" ? (
-                <NavItem
-                  id={`nav-link-${i}`}
-                  href={link.url}
-                  name={link.name}
-                  {...getActiveClassProps(link.url)}
-                  onClick={() => setNavOpen(false)}
-                />
-              ) : (
-                <Accordion
-                  triggerText={link.name}
-                  className={{
-                    trigger: "justify-center",
-                    content: "rounded border border-slate-400 my-2",
-                  }}
-                >
-                  <ol className="text-center">
-                    {link.children?.map((child: Link, i: number) => (
-                      <li key={i} className="my-4 px-2">
-                        <NavItem
-                          key={i}
-                          href={child.url}
-                          name={child.name}
-                          className="!inline-block text-black text-xl text-center rounded"
-                        />
-                      </li>
-                    ))}
-                  </ol>
-                </Accordion>
-              )}
-            </li>
-          ))}
-        </ol>
-      </>
-    );
-  }
-
-  // Normal navbar
   return (
-    <>
-      <ol className="hidden md:flex flex-row justify-center items-center">
-        {nav?.map((link: Link, i: number) => (
-          <li key={i} className="md:text-xl mx-2">
-            {link.type == "link" ? (
-              <NavItem
-                id={`nav-link-${i+1}`}
-                href={link.url}
-                name={link.name}
-                {...getActiveClassProps(link.url)}
-              />
-            ) : (
-              <NavigationMenu
-                link={link}
-                className={`${getActiveClassProps(link.url)?.className} text-xl`}
-              />
-            )}
-          </li>
-        ))}
-      </ol>
-
-      <button
-        className="cursor-pointer z-20 pr-4 text-gray-500 md:hidden"
-        onClick={() => setNavOpen(!isNavOpen)}
-      >
-        <Icon type="fas-bars" className="w-8" altText="Open navigation menu" />
-      </button>
-    </>
+    <Navbar isMenuOpen={isMenuOpen} onMenuOpenChange={setIsMenuOpen}>
+      <NavbarBrand>
+        <Logo>
+          <a href="/" className="rounded px-2 grid grid-rows-2" role="link">
+            <span className="text-lg md:text-2xl">{title}</span>
+            <span className="text-sm md:text-lg">{description}</span>
+          </a>
+        </Logo>
+      
+      </NavbarBrand>
+      <Fragment>
+      <NavbarContent className="hidden sm:flex gap-4" justify="end">
+        {data?.map((link: LinkType, i: number) => {
+          let tempChildComponent: React.ReactElement | undefined = undefined;
+          if (link) {
+            const isActive: boolean = isCurrentPath(link.url);
+            const props = getActiveClassProps(isActive);
+            if (link.type == "link") {
+              tempChildComponent = (
+                <Link color="foreground" href={link.url} aria-current={isActive ? "page" : undefined} {...props}>
+                  {link.name}
+                </Link>
+              );
+            } else {
+              tempChildComponent = <Dropdown data={link} className={{ trigger: `${props.className} p-0 m-0` }} />;
+            }
+            return (
+              <NavbarItem key={i} isActive={isActive}>
+                {tempChildComponent}
+              </NavbarItem>
+            );
+          }
+        })}
+      </NavbarContent>
+      <NavbarMenuToggle aria-label={isMenuOpen ? "Close Menu" : "Open Menu"} className="sm:hidden h-10 w-auto" />
+      </Fragment>
+      <NavbarMenu className="my-10">
+      {data?.map((link: LinkType, i: number) => {
+          let tempChildComponent: React.ReactElement | undefined = undefined;
+          if (link) {
+            const isActive: boolean = isCurrentPath(link.url);
+            const props = getActiveClassProps(isActive);
+            if (link.type == "link") {
+              tempChildComponent = (
+                <Link color="foreground" href={link.url} aria-current={isActive ? "page" : undefined} {...props} onClick={() => setIsMenuOpen(false)}>
+                  {link.name}
+                </Link>
+              );
+            } else {
+              tempChildComponent = <Dropdown data={link} className={{ trigger: `${props.className} p-0 m-0` }} onClickFn={setIsMenuOpen}/>;
+            }
+            return (
+              <NavbarMenuItem key={i} isActive={isActive}>
+                {tempChildComponent}
+              </NavbarMenuItem>
+            );
+          }
+        })}
+      </NavbarMenu>
+    </Navbar>
   );
 }
+
